backend/requirements.txt
# Authentication & Security
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
psycopg-binary==3.3.2

# OpenAI
openai==1.50.0

# File Processing
PyPDF2==3.0.1
python-docx==1.1.2

# Utilities
python-dotenv==1.0.1
requests==2.32.0
pydantic==2.9.0